--- a/docs/make.jl
+++ b/docs/make.jl
@@ -1,5 +1,14 @@
 using Documenter, FiniteDifferences
 
+# Copy the README.
+repo_root = dirname(dirname(@__FILE__))
+cp(
+    joinpath(repo_root, "README.md"),
+    joinpath(repo_root, "docs", "src", "index.md");
+    # `index.md` will already exist, so we set `force=true` to overwrite it.
+    force=true
+)
+
 makedocs(
     modules=[FiniteDifferences],
     format=Documenter.HTML(prettyurls=get(ENV, "CI", nothing) == "true"),