--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1,3 +1,4 @@
+using BenchmarkTools
 using ChainRulesCore
 using FiniteDifferences
 using LinearAlgebra