`master`:

```julia
julia> m = central_fdm(6, 1); @benchmark $m(sin, 1)
BenchmarkTools.Trial:
  memory estimate:  1.02 KiB
  allocs estimate:  56
  --------------
  minimum time:     2.220 μs (0.00% GC)
  median time:      2.307 μs (0.00% GC)
  mean time:        2.491 μs (1.43% GC)
  maximum time:     182.601 μs (98.30% GC)
  --------------
  samples:          10000
  evals/sample:     9
```

This PR:

```julia
julia> m = central_fdm(6, 1); @benchmark $m(sin, 1)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     451.449 ns (0.00% GC)
  median time:      466.278 ns (0.00% GC)
  mean time:        479.135 ns (0.00% GC)
  maximum time:     1.387 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     198
```

The current implementation is suboptimal in several ways:

1. The function `f` is evaluated at `x` and possibly again around `x` [to estimate the round-off error](https://github.com/JuliaDiff/FiniteDifferences.jl/blob/master/src/methods.jl#L29). This can be avoided by reusing the function evaluations of the bound estimator.

2. The bound estimator [may be run twice](https://github.com/JuliaDiff/FiniteDifferences.jl/blob/master/src/methods.jl#L361) if it returns zero identically, which requires even more function evaluations. This can be avoided by increasing the order of the bound estimator by one so that it can estimate the derivative in a neighbourhood of `x` instead of just at `x`.

3. The functions `estimate_magnitude` and `estimate_roundoff_error`, whilst fine in principle, feel a little ad hoc and complicate the logic.

This PR gets rid of `estimate_magnitude` and `estimate_roundoff_error` entirely by more cleverly using the function evaluations of the bound estimator (`_estimate_magnitudes`). Moreover, the PR makes sure that unnecessary allocations are avoided where possible.

Improvements:
- [x] Reduce number of allocations.
- [x] Reduce number of function evaluations and ensure that the number of function evaluations is always the same.
- [x] Improve robustness of adaptation.
- [x] Improve mechanism for step size capping.
- [x] Eliminate dependence on arbitrary thresholds and numbers.
- [x] Allow non-floats in`extrapolate_fdm`.
- [x] Automatically copy `README.md` to `docs/src/index.md` in `docs/make.jl`.

Additions:
- [x] More elaborate accuracy tests.
- [x] Expansion of README. 
- [x] Resolves #129: Add a `max_range` keyword that specifies how far `f` is allowed to be evaluated from `x`:
```julia
julia> m = central_fdm(5, 1, max_range=1e-6); grad(m, sum, ones(5))
```

Breaking changes:
- [x] Remove `max_step` keyword argument.