@oxinabox 2021-01-02
Maybe StaticArrays.jl, rather than tuples?