@wesselb 2021-01-04
With the current spaghetti code, allocations are down to zero, which achieves a 4x speed-up compared to `master`:

```julia
julia> using FiniteDifferences, BenchmarkTools

julia> m = central_fdm(10, 2, adapt=2)
@benchmark m(sinFiniteDifferenceMethod:
  order of method:       10
  order of derivative:   2
  grid:                  [-5, -4, -3, -2, -1, 1, 2, 3, 4, 5]
  coefficients:          [-0.011298500881834215, 0.11119929453262786, -0.4830357142857143, 1.1558201058201059, -0.7726851851851851, -0.7726851851851851, 1.1558201058201059, -0.4830357142857143, 0.11119929453262786, -0.011298500881834215]

julia> m(sin, 1) + sin(1)
3.8624659026709196e-13

julia> @benchmark $m(sin, 1)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     916.167 ns (0.00% GC)
  median time:      926.521 ns (0.00% GC)
  mean time:        1.027 μs (0.00% GC)
  maximum time:     3.898 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     48
```

`master`:
```julia
julia> m(sin, 1) + sin(1)
2.97983859809392e-13

julia> @benchmark $m(sin, 1)
BenchmarkTools.Trial:
  memory estimate:  1.80 KiB
  allocs estimate:  92
  --------------
  minimum time:     3.824 μs (0.00% GC)
  median time:      3.942 μs (0.00% GC)
  mean time:        4.196 μs (1.39% GC)
  maximum time:     207.103 μs (97.42% GC)
  --------------
  samples:          10000
  evals/sample:     8
```