@oxinabox 2021-01-08
we need to drop Julia-Nightly from the CI (out of scope for this PR)