@wesselb 2021-01-10
@oxinabox The ChainRules integration test fails because ChainRules and ChainRulesUtils both have a `[compat]` entry that does not allow version 0.12.0 for FiniteDifferences. I tried to fix this by removing the entry from the Project.toml in the cloned ChainRules repo in the GitHub workflow, but that fails because also ChainRulesUtils restricts FiniteDifferences. I there a way to tell Julia to ignore requirements for FiniteDifferences?