@wesselb 2021-01-10
Fair enough, I guess we can just check this PR locally. Running the ChainRules tests locally, it appears that some tests fail due to singularities, but those failures are resolved by changing

```julia
const _fdm = central(5, 1)
```
to

```julia
const _fdm = central(5, 1; max_range=1e-2)
```

in `src/ChainRulesTestUtils.jl`. I'll make a PR.

Are we happy to merge this?